@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2013 University of Stuttgart.
+ * Copyright (c) 2013-2017 University of Stuttgart.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * and the Apache License 2.0 which both accompany this distribution,
@@ -7,7 +7,7 @@
  * and http://www.apache.org/licenses/LICENSE-2.0
  *
  * Contributors:
- *     Oliver Kopp - initial API and implementation
+ *     Oliver Kopp - initial API and implementation, merge with org.eclipse.winery.repository.datatypes.ids.IdNames
  *******************************************************************************/
 package org.eclipse.winery.common.ids;
 
@@ -28,4 +28,15 @@ public class IdNames {
 	public static final String SOURCEINTERFACES = "sourceinterfaces";
 	public static final String TARGETINTERFACES = "targetinterfaces";
 	public static final String TOPOLOGYTEMPATE = "topologytemplate";
+
+	// the files belonging to one artifact template are nested in the sub
+	// directory "files"
+	public static final String ARTIFACTTEMPLATEFILESDIRECTORY = "files";
+	public static final String ARTIFACTTEMPLATESRCDIRECTORY = "src";
+	
+	public static final String CONSTRAINTTYPES = "constrainttypes";
+	public static final String NAMESPACES = "namespaces";
+	public static final String PLANLANGUAGES = "planlanguages";
+	public static final String PLANTYPES = "plantypes";
+
 }
