@@ -86,7 +86,7 @@ import org.eclipse.jdt.annotation.Nullable;
         TArtifactType.class,
         TPolicyType.class
 })
-public class TEntityType extends TExtensibleElements {
+public class TEntityType extends TExtensibleElements implements HasName, HasInheritance, HasTargetNamespace {
     @XmlElement(name = "Tags")
     protected TTags tags;
     @XmlElement(name = "DerivedFrom")
@@ -195,30 +195,17 @@ public class TEntityType extends TExtensibleElements {
         this.propertiesDefinition = value;
     }
 
-    /**
-     * Gets the value of the name property.
-     *
-     * @return possible object is {@link String }
-     */
     @NonNull
+    @Override
     public String getName() {
         return name;
     }
 
-    /**
-     * Sets the value of the name property.
-     *
-     * @param value allowed object is {@link String }
-     */
+    @Override
     public void setName(String value) {
         this.name = value;
     }
 
-    /**
-     * Gets the value of the abstract property.
-     *
-     * @return possible object is {@link TBoolean }
-     */
     @NonNull
     public TBoolean getAbstract() {
         if (_abstract == null) {
@@ -297,7 +284,7 @@ public class TEntityType extends TExtensibleElements {
      */
     @XmlAccessorType(XmlAccessType.FIELD)
     @XmlType(name = "")
-    public static class DerivedFrom {
+    public static class DerivedFrom implements HasType {
 
         @XmlAttribute(name = "typeRef", required = true)
         protected QName typeRef;
@@ -319,6 +306,20 @@ public class TEntityType extends TExtensibleElements {
          */
         public void setTypeRef(QName value) {
             this.typeRef = value;
+        }
+
+        public QName getType() {
+            return this.getTypeRef();
+        }
+
+        @Override
+        public QName getTypeAsQName() {
+            return this.getType();
+        }
+
+        @Override
+        public void setType(QName type) {
+            this.setTypeRef(type);
         }
     }
 
