@@ -8,7 +8,9 @@
  *
  * Contributors:
  *     Niko Stadelmaier - initial API and implementation
+ *     Lukas Balzer - corrected image upload of SelfServicePortal
  */
+
 import { Injectable } from '@angular/core';
 import { Router } from '@angular/router';
 import { Headers, Http, RequestOptions } from '@angular/http';
@@ -42,32 +44,44 @@ export interface SelfServiceApiData {
 @Injectable()
 export class SelfServicePortalService {
 
+    private url: string;
     private path: string;
     selfServiceData: SelfServiceApiData;
 
     constructor(private http: Http,
                 private route: Router) {
+        let path = this.route.url;
+        path = path.substring(0, path.lastIndexOf('/'));
+        this.url = decodeURIComponent(path);
+    }
+
+    getIconPath(): string {
+        return backendBaseURL + this.url + '/icon.jpg';
+    }
+
+    getImagePath(): string {
+        return backendBaseURL + this.url + '/image.jpg';
     }
 
     getSelfServiceData(): Observable<SelfServiceApiData> {
-        const headers = new Headers({'Accept': 'application/json'});
-        const options = new RequestOptions({headers: headers});
+        const headers = new Headers({ 'Accept': 'application/json' });
+        const options = new RequestOptions({ headers: headers });
 
         return this.http.get(backendBaseURL + this.path, options)
             .map(res => this.selfServiceData = res.json());
     }
 
     saveName(displayName: string): Observable<any> {
-        return this.saveSingleProperty({'displayName': displayName}, this.path + 'displayname');
+        return this.saveSingleProperty({ 'displayName': displayName }, this.path + 'displayname');
     }
 
     saveDescription(description: string): Observable<any> {
-        return this.saveSingleProperty({'description': description}, this.path + 'description');
+        return this.saveSingleProperty({ 'description': description }, this.path + 'description');
     }
 
     saveSingleProperty(property: any, path: string): Observable<any> {
-        const headers = new Headers({'Content-Type': 'application/json'});
-        const options = new RequestOptions({headers: headers});
+        const headers = new Headers({ 'Content-Type': 'application/json' });
+        const options = new RequestOptions({ headers: headers });
 
         return this.http.put(backendBaseURL + path, JSON.stringify(property), options);
     }
