@@ -11,15 +11,18 @@
  *******************************************************************************/
 package org.eclipse.winery.repository.backend.constants;
 
-import javax.ws.rs.core.MediaType;
-
 import org.eclipse.winery.common.constants.MimeTypes;
+
+import org.apache.tika.mime.MediaType;
 
 /**
  * see also {@link org.eclipse.winery.common.constants.MimeTypes}
  */
 public class MediaTypes {
 
-	public static final MediaType MEDIATYPE_TOSCA_DEFINITIONS = MediaType.valueOf(MimeTypes.MIMETYPE_TOSCA_DEFINITIONS);
+	public static final MediaType MEDIATYPE_TOSCA_DEFINITIONS = MediaType.parse(MimeTypes.MIMETYPE_TOSCA_DEFINITIONS);
+	public static final MediaType MEDIATYPE_APPLICATION_JSON = MediaType.parse("application/json");
+	public static final MediaType MEDIATYPE_TEXT_XML = MediaType.parse("text/xml");
+	public static final MediaType MEDIATYPE_XSD = MediaType.parse(MimeTypes.MIMETYPE_XSD);
 
 }
