@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2012-2016 University of Stuttgart.
+ * Copyright (c) 2012-2017 University of Stuttgart.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * and the Apache License 2.0 which both accompany this distribution,
@@ -7,10 +7,11 @@
  * and http://www.apache.org/licenses/LICENSE-2.0
  *
  * Contributors:
- *     Oliver Kopp - initial API and implementation
+ *     Oliver Kopp - initial API and implementation, adaptions to IGenericRepository
  *     Tino Stadelmaier, Philipp Meyer - rename id and/or namespace
  *     Lukas Harzentter - get namespaces for specific component
  *     Nicole Keppler - forceDelete for Namespaces
+ *     Philipp Meyer - support for source directory
  *******************************************************************************/
 package org.eclipse.winery.repository.backend.filebased;
 
@@ -45,13 +46,12 @@ import java.util.zip.ZipEntry;
 import java.util.zip.ZipInputStream;
 import java.util.zip.ZipOutputStream;
 
-import javax.ws.rs.core.MediaType;
-
 import org.eclipse.winery.common.RepositoryFileReference;
 import org.eclipse.winery.common.Util;
 import org.eclipse.winery.common.ids.GenericId;
 import org.eclipse.winery.common.ids.Namespace;
 import org.eclipse.winery.common.ids.XMLId;
+import org.eclipse.winery.common.ids.admin.NamespacesId;
 import org.eclipse.winery.common.ids.definitions.ArtifactTemplateId;
 import org.eclipse.winery.common.ids.definitions.ArtifactTypeId;
 import org.eclipse.winery.common.ids.definitions.CapabilityTypeId;
@@ -66,26 +66,36 @@ import org.eclipse.winery.common.ids.definitions.ServiceTemplateId;
 import org.eclipse.winery.common.ids.definitions.TOSCAComponentId;
 import org.eclipse.winery.common.ids.elements.TOSCAElementId;
 import org.eclipse.winery.model.tosca.Definitions;
+import org.eclipse.winery.model.tosca.HasIdInIdOrNameField;
 import org.eclipse.winery.repository.Constants;
 import org.eclipse.winery.repository.backend.AbstractRepository;
 import org.eclipse.winery.repository.backend.BackendUtils;
 import org.eclipse.winery.repository.backend.IRepositoryAdministration;
+import org.eclipse.winery.repository.backend.NamespaceManager;
+import org.eclipse.winery.repository.backend.RepositoryFactory;
 import org.eclipse.winery.repository.backend.constants.MediaTypes;
-import org.eclipse.winery.repository.resources.AbstractComponentInstanceResource;
-import org.eclipse.winery.repository.resources.AbstractComponentsResource;
+import org.eclipse.winery.repository.backend.xsd.RepositoryBasedXsdImportManager;
+import org.eclipse.winery.repository.backend.xsd.XsdImportManager;
+import org.eclipse.winery.repository.configuration.FileBasedRepositoryConfiguration;
+import org.eclipse.winery.repository.exceptions.WineryRepositoryException;
 
+import org.apache.commons.compress.archivers.ArchiveException;
+import org.apache.commons.compress.archivers.ArchiveOutputStream;
+import org.apache.commons.compress.archivers.ArchiveStreamFactory;
+import org.apache.commons.compress.archivers.zip.ZipArchiveEntry;
 import org.apache.commons.configuration.Configuration;
 import org.apache.commons.configuration.ConfigurationException;
 import org.apache.commons.configuration.PropertiesConfiguration;
+import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang3.SystemUtils;
+import org.apache.tika.mime.MediaType;
 import org.eclipse.jgit.dircache.InvalidPathException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
- * When it comes to a storage of plain files, we use Java 7's nio internally.
- * Therefore, we intend to expose the stream types offered by java.nio.Files:
- * BufferedReader/BufferedWriter
+ * When it comes to a storage of plain files, we use Java 7's nio internally. Therefore, we intend to expose the stream
+ * types offered by java.nio.Files: BufferedReader/BufferedWriter
  */
 public class FilebasedRepository extends AbstractRepository implements IRepositoryAdministration {
 
@@ -99,12 +109,12 @@ public class FilebasedRepository extends AbstractRepository implements IReposito
 	private final FileSystemProvider provider;
 
 	/**
-	 *
-	 * @param repositoryLocation a string pointing to a location on the file
-	 *            system. May be null.
+	 * @param fileBasedRepositoryConfiguration configuration of the filebased repository. The contained repositoryPath
+	 *                                         may be null
 	 */
-	public FilebasedRepository(String repositoryLocation) {
-		this.repositoryRoot = this.determineRepositoryPath(repositoryLocation);
+	public FilebasedRepository(FileBasedRepositoryConfiguration fileBasedRepositoryConfiguration) {
+		Objects.requireNonNull(fileBasedRepositoryConfiguration);
+		this.repositoryRoot = this.determineRepositoryPath(fileBasedRepositoryConfiguration.getRepositoryPath());
 		this.fileSystem = this.repositoryRoot.getFileSystem();
 		this.provider = this.fileSystem.provider();
 		LOGGER.debug("Repository root: {}", this.repositoryRoot);
@@ -127,31 +137,23 @@ public class FilebasedRepository extends AbstractRepository implements IReposito
 	}
 
 	private Path id2AbsolutePath(GenericId id) {
-		Path relativePath = this.fileSystem.getPath(BackendUtils.getPathInsideRepo(id));
+		Path relativePath = this.fileSystem.getPath(Util.getPathInsideRepo(id));
 		return this.makeAbsolute(relativePath);
 	}
 
 	/**
-	 * Converts the given reference to an absolute path of the underlying
-	 * FileSystem
+	 * Converts the given reference to an absolute path of the underlying FileSystem
 	 */
 	public Path ref2AbsolutePath(RepositoryFileReference ref) {
 		return this.id2AbsolutePath(ref.getParent()).resolve(ref.getFileName());
 	}
 
-	protected Path determineRepositoryPath(String repositoryLocation) {
+	protected Path determineRepositoryPath(final Path configuredRepositoryPath) {
 		Path repositoryPath;
-		if (repositoryLocation == null) {
+		if (configuredRepositoryPath == null) {
 			if (SystemUtils.IS_OS_WINDOWS) {
-				if (new File(Constants.GLOBAL_REPO_PATH_WINDOWS).exists()) {
-					repositoryLocation = Constants.GLOBAL_REPO_PATH_WINDOWS;
-					File repo = new File(repositoryLocation);
-					try {
-						org.apache.commons.io.FileUtils.forceMkdir(repo);
-					} catch (IOException e) {
-						FilebasedRepository.LOGGER.error("Could not create repository directory", e);
-					}
-					repositoryPath = repo.toPath();
+				if (Files.exists(Constants.GLOBAL_REPO_PATH_WINDOWS)) {
+					repositoryPath = Constants.GLOBAL_REPO_PATH_WINDOWS;
 				} else {
 					repositoryPath = this.createDefaultRepositoryPath();
 				}
@@ -159,13 +161,12 @@ public class FilebasedRepository extends AbstractRepository implements IReposito
 				repositoryPath = this.createDefaultRepositoryPath();
 			}
 		} else {
-			File repo = new File(repositoryLocation);
 			try {
-				org.apache.commons.io.FileUtils.forceMkdir(repo);
+				org.apache.commons.io.FileUtils.forceMkdir(configuredRepositoryPath.toFile());
 			} catch (IOException e) {
 				FilebasedRepository.LOGGER.error("Could not create repository directory", e);
 			}
-			repositoryPath = repo.toPath();
+			repositoryPath = configuredRepositoryPath;
 		}
 		return repositoryPath;
 	}
@@ -241,8 +242,7 @@ public class FilebasedRepository extends AbstractRepository implements IReposito
 			return;
 		}
 
-		AbstractComponentInstanceResource componentInstanceResource = AbstractComponentsResource.getComponentInstaceResource(oldId);
-		//AbstractComponentInstanceResource newComponentInstanceResource = AbstractComponentsResource.getComponentInstaceResource(newId);
+		Definitions definitions = this.getDefinitions(oldId);
 
 		RepositoryFileReference oldRef = BackendUtils.getRefOfDefinitions(oldId);
 		RepositoryFileReference newRef = BackendUtils.getRefOfDefinitions(newId);
@@ -256,23 +256,22 @@ public class FilebasedRepository extends AbstractRepository implements IReposito
 		org.apache.commons.io.FileUtils.moveDirectory(oldDir, newDir);
 
 		// Update definitions and store it
-		Definitions definitions = componentInstanceResource.getDefinitions();
 
 		// This also updates the definitions of componentInstanceResource
 		BackendUtils.updateWrapperDefinitions(newId, definitions);
 
 		// This works, because the definitions object here is the same as the definitions object treated at copyIdToFields
 		// newId has to be passed, because the id is final at AbstractComponentInstanceResource
-		componentInstanceResource.copyIdToFields(newId);
+		BackendUtils.copyIdToFields((HasIdInIdOrNameField) definitions.getElement(), newId);
 
 		try {
 			BackendUtils.persist(definitions, newRef, MediaTypes.MEDIATYPE_TOSCA_DEFINITIONS);
 		} catch (InvalidPathException e) {
+			LOGGER.debug("Invalid path during write", e);
 			// QUICK FIX
 			// Somewhere, the first letter is deleted --> /odetypes/http%3A%2F%2Fwww.example.org%2F05/
 			// We just ignore it for now
 		}
-
 	}
 
 	public void forceDelete(Class<? extends TOSCAComponentId> toscaComponentIdClazz, Namespace namespace) {
@@ -294,9 +293,6 @@ public class FilebasedRepository extends AbstractRepository implements IReposito
 		return Files.exists(absolutePath);
 	}
 
-	/**
-	 * {@inheritDoc}
-	 */
 	@Override
 	public void putContentToFile(RepositoryFileReference ref, String content, MediaType mediaType) throws IOException {
 		if (mediaType == null) {
@@ -311,9 +307,6 @@ public class FilebasedRepository extends AbstractRepository implements IReposito
 		Files.write(path, content.getBytes());
 	}
 
-	/**
-	 * {@inheritDoc}
-	 */
 	@Override
 	public void putContentToFile(RepositoryFileReference ref, InputStream inputStream, MediaType mediaType) throws IOException {
 		if (mediaType == null) {
@@ -510,17 +503,17 @@ public class FilebasedRepository extends AbstractRepository implements IReposito
 		// @formatter:off
 		@SuppressWarnings("rawtypes")
 		Collection<Class<? extends TOSCAComponentId>> toscaComponentIds = Arrays.asList(
-			ArtifactTemplateId.class,
-			ArtifactTypeId.class,
-			CapabilityTypeId.class,
-			NodeTypeId.class,
-			NodeTypeImplementationId.class,
-			PolicyTemplateId.class,
-			PolicyTypeId.class,
-			RelationshipTypeId.class,
-			RelationshipTypeImplementationId.class,
-			RequirementTypeId.class,
-			ServiceTemplateId.class
+				ArtifactTemplateId.class,
+				ArtifactTypeId.class,
+				CapabilityTypeId.class,
+				NodeTypeId.class,
+				NodeTypeImplementationId.class,
+				PolicyTemplateId.class,
+				PolicyTypeId.class,
+				RelationshipTypeId.class,
+				RelationshipTypeImplementationId.class,
+				RequirementTypeId.class,
+				ServiceTemplateId.class
 		);
 		// @formatter:on
 
@@ -532,6 +525,16 @@ public class FilebasedRepository extends AbstractRepository implements IReposito
 		Collection<Class<? extends TOSCAComponentId>> list = new ArrayList<>();
 		list.add(clazz);
 		return getNamespaces(list);
+	}
+
+	@Override
+	public NamespaceManager getNamespaceManager() {
+		return new ConfigurationBasedNamespaceManager(this.getConfiguration(new NamespacesId()));
+	}
+
+	@Override
+	public XsdImportManager getXsdImportManager() {
+		return new RepositoryBasedXsdImportManager();
 	}
 
 	private Collection<Namespace> getNamespaces(Collection<Class<? extends TOSCAComponentId>> toscaComponentIds) {
@@ -632,30 +635,42 @@ public class FilebasedRepository extends AbstractRepository implements IReposito
 		}
 	}
 
-	/**
-	 * {@inheritDoc}
-	 */
 	@Override
 	public long getSize(RepositoryFileReference ref) throws IOException {
 		return Files.size(this.ref2AbsolutePath(ref));
 	}
 
-	/**
-	 * {@inheritDoc}
-	 */
 	@Override
 	public FileTime getLastModifiedTime(RepositoryFileReference ref) throws IOException {
 		Path path = this.ref2AbsolutePath(ref);
 		return Files.getLastModifiedTime(path);
 	}
 
-	/**
-	 * {@inheritDoc}
-	 */
 	@Override
 	public InputStream newInputStream(RepositoryFileReference ref) throws IOException {
 		Path path = this.ref2AbsolutePath(ref);
 		return Files.newInputStream(path);
 	}
 
+	@Override
+	public void getZippedContents(final GenericId id, OutputStream out) throws WineryRepositoryException {
+		Objects.requireNonNull(id);
+		Objects.requireNonNull(out);
+
+		SortedSet<RepositoryFileReference> containedFiles = this.getContainedFiles(id);
+
+		try (final ArchiveOutputStream zos = new ArchiveStreamFactory().createArchiveOutputStream("zip", out)) {
+			for (RepositoryFileReference ref : containedFiles) {
+				zos.putArchiveEntry(new ZipArchiveEntry(ref.getFileName()));
+				try (InputStream is = RepositoryFactory.getRepository().newInputStream(ref)) {
+					IOUtils.copy(is, zos);
+				}
+				zos.closeArchiveEntry();
+			}
+		} catch (ArchiveException e) {
+			throw new WineryRepositoryException("Internal error while generating archive", e);
+		} catch (IOException e) {
+			throw new WineryRepositoryException("I/O exception during export", e);
+		}
+	}
 }
