@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2012-2013 University of Stuttgart.
+ * Copyright (c) 2012-2017 University of Stuttgart.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * and the Apache License 2.0 which both accompany this distribution,
@@ -8,21 +8,18 @@
  *
  * Contributors:
  *     Oliver Kopp - initial API and implementation
+ *     Philipp Meyer - support for source directory
  *******************************************************************************/
 package org.eclipse.winery.repository.datatypes.ids.elements;
 
 import org.eclipse.winery.common.ids.XMLId;
 import org.eclipse.winery.common.ids.definitions.ArtifactTemplateId;
 import org.eclipse.winery.common.ids.elements.TOSCAElementId;
-import org.eclipse.winery.repository.datatypes.ids.IdNames;
 
 public class ArtifactTemplateDirectoryId extends TOSCAElementId {
 
-	private final static XMLId xmlID = new XMLId(IdNames.ARTIFACTTEMPLATEDIRECTORY, false);
 
-
-	public ArtifactTemplateDirectoryId(ArtifactTemplateId parent) {
-		super(parent, ArtifactTemplateDirectoryId.xmlID);
+	public ArtifactTemplateDirectoryId(ArtifactTemplateId parent, String directoryId) {
+		super(parent, new XMLId(directoryId, false));
 	}
-
 }
