@@ -15,7 +15,7 @@ import { isNullOrUndefined } from 'util';
 
 import { Node } from '../model/workflow/node';
 import { BroadcastService } from './broadcast.service';
-import {NodeService} from './node.service';
+import { NodeService } from './node.service';
 
 /**
  * ModelService
@@ -70,8 +70,6 @@ export class ModelService {
     }
 
     public save() {
-        console.log('****************** save data *********************');
-        console.log(this.nodes);
         this.broadcastService.broadcast(this.broadcastService.saveEvent, JSON.stringify(this.nodes));
     }
 
