@@ -8,17 +8,16 @@
  *
  * Contributors:
  *     ZTE - initial API and implementation and/or initial documentation
+ *     Lukas Harzenetter - fix usage of angular-cli
  */
-
 import { enableProdMode } from '@angular/core';
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 import { AppModule } from './app/app.module';
+import { environment } from '../environments/environment';
 
-if (process.env.ENV === 'production') {
+if (environment.production) {
     enableProdMode();
 }
 // noinspection TypeScriptValidateTypes
 platformBrowserDynamic().bootstrapModule(AppModule);
 
-// global styles
-import './index.css';
