@@ -8,9 +8,9 @@
  *
  * Contributors:
  *     ZTE - initial API and implementation and/or initial documentation
+ *
+ * This file has been generated using `ng new` and adapted to winery's needs.
  */
-// This file is required by karma.conf.js and loads recursively all the .spec and framework files
-
 import { getTestBed } from '@angular/core/testing';
 import {
   BrowserDynamicTestingModule,
@@ -23,19 +23,15 @@ import 'zone.js/dist/long-stack-trace-zone';
 import 'zone.js/dist/proxy.js';
 import 'zone.js/dist/sync-test';
 
-// Unfortunately there's no typing for the `__karma__` variable. Just declare it as any.
 declare var __karma__: any;
 declare var require: any;
 
-// Prevent Karma from running prematurely.
 __karma__.loaded = () => {};
 
-// First, initialize the Angular testing environment.
 getTestBed().initTestEnvironment( BrowserDynamicTestingModule,
   platformBrowserDynamicTesting());
-// Then we find all the tests.
+
 const context = require.context('./', true, /\.spec\.ts$/);
-// And load the modules.
 context.keys().map(context);
-// Finally, start Karma to run the tests.
+
 __karma__.start();
